# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/quest/{quest_id}/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get all teams by quest id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quest id",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.Team"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Create new team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quest ID",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Desired team information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/teams.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                }
            }
        },
        "/teams/join/{invite_path}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Join team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team invite url param",
                        "name": "invite_path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                }
            }
        },
        "/teams/{team_id}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get team by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Change team information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/teams.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Delete team by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/teams/{team_id}/captain": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Change team captain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Change captain request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/teams.ChangeLeaderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/teams/{team_id}/leave": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Leave the team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New captain (if leader leaves)",
                        "name": "new_captain_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/teams/{team_id}/{member_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Remove member from team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Member id",
                        "name": "member_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "storage.Team": {
            "type": "object",
            "properties": {
                "captain": {
                    "$ref": "#/definitions/storage.User"
                },
                "id": {
                    "type": "string"
                },
                "invite_link": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.User"
                    }
                },
                "name": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
...
        "storage.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "teams.ChangeLeaderRequest": {
            "type": "object",
            "properties": {
                "new_captain_id": {
                    "type": "string"
                }
            }
        },
        "teams.CreateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "teams.UpdateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}